backend/patient_service/requirements.txt
# Python dependencies (FastAPI, SQLAlchemy, etc.). 
# Dependencies for this service only  

fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt==4.0.1]
pydantic[email]
python-jose[cryptography]
pytest
# PyJWT
# jwt
pydantic-settings

